date: 19:20 2017/12/3

+++ src/client/BL/Accountant/FinancialAccountbl/FinancialAccountController.java
@@ -73,7 +73,7 @@ public class FinancialAccountController implements FinancialAccountInterface {
 
         String id=co.getKeyno();
         String name= co.getKeyname();
-        Long money = co.getSumall();
+        Double money = co.getSumall();
         Account account = new Account(id,name,money.toString());
         return account;
     }
@@ -82,7 +82,7 @@ public class FinancialAccountController implements FinancialAccountInterface {
 
         String id=account.getaccountID();
         String name= account.getaccountName();
-        Long money =Long.parseLong(account.getmoney());
+        Double money = Double.parseDouble(account.getmoney());
         coVO co = new coVO();
         co.setKeyname(name);
         co.setSumall(money);
