+++ src/client/Vo/coVO.java
@@ -4,11 +4,11 @@ import java.io.Serializable;
 
 public class coVO implements Serializable {
 	private String keyname;
-	private Long sumall;
+	private Double sumall;
 	private String keyno;
 
 
-	public coVO(String keyname, Long sumall, String keyno) {
+	public coVO(String keyname, Double sumall, String keyno) {
 		this.keyname = keyname;
 		this.sumall = sumall;
 		this.keyno = keyno;
@@ -26,11 +26,11 @@ public class coVO implements Serializable {
 		this.keyname = keyname;
 	}
 
-	public Long getSumall() {
+	public Double getSumall() {
 		return sumall;
 	}
 
-	public void setSumall(Long sumall) {
+	public void setSumall(Double sumall) {
 		this.sumall = sumall;
 	}
